# json: extendSlice optimization is still valid for go1.19