--- a/json/reflect.go
+++ b/json/reflect.go
@@ -1,5 +1,5 @@
-//go:build go1.18
-// +build go1.18
+//go:build go1.20
+// +build go1.20
 
 package json