@achille-roussel 2022-10-07
It's been 5 versions of Go since we introduced this optimization and it's never been invalidated by the runtime. Maybe we would spend less engineering time if we were to remove the build tag and address if the build ever breaks in a future Go version?